Revert "Fixes some issues parsing gitweb, should make it work on 1.7.9"
This reverts commit d953a6088bd67c42171e5fdbea56bdc485d1c480
which breaks parsing of commit message and changed files list on Linux box.

On Linux box PHP_EOL is equal to "\n" and on Windows box it is equal to "\r\n".
Also, Gitweb on Windows box returns EOL as "\r\n",
on Linux - "\n". So, cross-linking (Gitweb on Windows and
Mantis SCM plugin on Linux, and vice versa) was impossible.

Now it is obviuos to remove "Line feeds" and "Carrage returns"
separately to make such cross-linking possible.